refactor: improve inter-thread communication logic
refactor:
* Update logic of event loops in method animate and function
    kinematics for an improved efficiency of communication between their
    associated threads.
- Eliminate any unsigned short since there was no real need for them and
    their removal improves the efficiency of the code.
* Redefine certain long variables as short since there could easily be
    accommodated by the short type.

doc:
* Update some comments.
+ Add some comments.
- Eliminate excess comments.

Note on concurrency stability: This logic is considered stable, though
without locks. Mutexes and locks may improve efficiency.

+++ src/display.h
@@ -17,10 +17,10 @@ class Display {
 
   public:
     bool idle{true};
-    const unsigned short screen_width{480};
-    const unsigned short screen_height{480};
-    const long frame_life{41}; // millisec, so that frame_rate ~= 24 Hz
-
+    const short screen_width{480};
+    const short screen_height{480};
+    const short frame_life{41};
+      // frame_life is 41 millisec, so that frame_rate ~= 24 Hz
     void animate();
 };
